Table log as new component and some fix
