+++ js/extension/components/tabou2IdentifyPanel/modals/Tabou2LogsModal.jsx
@@ -1,16 +1,9 @@
 import React, {useState, useEffect, useRef} from 'react';
 import ResizableModal from '@mapstore/components/misc/ResizableModal';
-import { Grid, Col, Table, Glyphicon, Row } from 'react-bootstrap';
-import { isEqual, orderBy, find, isEmpty, maxBy, omit } from 'lodash';
-import Tabou2Combo from '@ext/components/form/Tabou2Combo';
-import Tabou2TextForm from '@ext/components/form/Tabou2TextForm';
+import { isEqual, find, maxBy, omit } from 'lodash';
 import { LOG_SCHEMA } from '@ext/constants';
-import { getTypesEvents } from "@ext/api/search";
-import ButtonRB from '@mapstore/components/misc/Button';
-import tooltip from '@mapstore/components/misc/enhancers/tooltip';
 import Message from "@mapstore/components/I18N/Message";
-
-const Button = tooltip(ButtonRB);
+import Tabou2LogsTable from './Tabou2LogsTable';
 
 export default function Tabou2LogsModal({
     visible,
@@ -45,12 +38,11 @@ export default function Tabou2LogsModal({
     // save log
     const saveEvent = (log) => {
         // call action to add log
-        let logToSend = omit(log, ["new", "id"]);
         if (!props.events.map(e => e.id).includes(log.id)) {
             // new
-            props.addEvent(logToSend);
+            props.addEvent(omit(log, ["new", "edit", "id"]));
         } else {
-            props.changeEvent(logToSend);
+            props.changeEvent(omit(log, ["new", "edit"]));
         }
         disabledAdd.current = false;
         editionActivate.current = false;
@@ -80,7 +72,7 @@ export default function Tabou2LogsModal({
         setLogs([...logs.filter(lo => lo.id !== log.id)]);
         editionActivate.current = false;
         disabledAdd.current = false;
-        props.deleteEvent(log);
+        if (!log.new) props.deleteEvent(log);
     };
 
     // Manage sort behavior
@@ -91,21 +83,6 @@ export default function Tabou2LogsModal({
             order = sortField[1][idx] === "asc" ? "desc" : "asc";
         }
         setSortField([[name, "id"], [order, "asc"]]);
-    };
-
-    // Manage sort icon
-    const getSortIcon = (name) => {
-        let icon = "sort-by-attributes";
-        let idx = sortField[0].indexOf(name);
-        if (idx > -1) {
-            icon = sortField[1][idx] === "asc" ? "sort-by-attributes" : "sort-by-attributes-alt";
-        }
-        return (<Glyphicon onClick={() => changeSort(name) } glyph={icon} style={{marginLeft: "5px"}}/>);
-    };
-
-    // TODO : fix this
-    const getStyle = (name) => {
-        return find(sortField, [0, name]) ? {color: "darkcyan"} : {color: "rgb(51, 51, 51)"};
     };
 
     // toolbar button
@@ -133,151 +110,19 @@ export default function Tabou2LogsModal({
             buttons={readOnly ? [] : buttons}
             onClose={onClick}
             size="lg">
-            <Grid fluid style={{overflow: "auto", height: "100%"}}>
-                <Row>
-                    <Col xs={12}>
-                        <Table>
-                            <thead>
-                                <tr>
-                                    <th className="col-xs-2" style={getStyle("eventDate")}><Message msgId="tabou2.logsModal.date"/>
-                                        {
-                                            getSortIcon("eventDate")
-                                        }
-                                    </th>
-                                    <th className="col-xs-2" style={getStyle("modifUser")}><Message msgId="tabou2.logsModal.owner"/>
-                                        {
-                                            getSortIcon("modifUser")
-                                        }
-                                    </th>
-                                    <th className="col-xs-2" style={getStyle("idType")}><Message msgId="tabou2.logsModal.type"/>
-                                    </th>
-                                    <th><Message msgId="tabou2.logsModal.editor"/></th>
-                                    <th><Message msgId="tabou2.logsModal.note"/></th>
-                                    {
-                                        readOnly ? null : (<th><Message msgId="tabou2.logsModal.actions"/></th>)
-                                    }
-                                </tr>
-                            </thead>
-                            <tbody style={{overflow: "auto"}}>
-                                {
-                                    orderBy(logs, sortField[0], sortField[1]).map((log, i) => (
-                                        <>
-                                            <tr>
-                                                <td>
-                                                    {
-                                                        new Date(log.eventDate).toLocaleDateString()
-                                                    }
-                                                </td>
-                                                <td>
-                                                    {
-                                                        !log.new ? log.createUser : props?.authent?.user || ""
-                                                    }
-                                                </td>
-                                                <td>
-                                                    {
-                                                        log.new || log.edit ? (
-                                                            <Tabou2Combo
-                                                                load={() => getTypesEvents()}
-                                                                valueField={"id"}
-                                                                defaultValue={log.edit ? log.typeEvenement.id : null}
-                                                                placeholder={props.i18n(props.messages, "tabou2.logsModal.typePlaceholder")}
-                                                                filter="contains"
-                                                                dropUp={i < 6 ? false : true}
-                                                                textField={"libelle"}
-                                                                onLoad={(r) => r?.elements || r}
-                                                                onSelect={(t) => {
-                                                                    changeLog({...log, typeEvenement: t});
-                                                                }}
-                                                                onChange={(t) => !t ? changeLog({...log, typeEvenement: null}) : null}
-                                                                messages={{
-                                                                    emptyList: props.i18n(props.messages, "tabou2.emptyList"),
-                                                                    openCombobox: props.i18n(props.messages, "tabou2.displaylist")
-                                                                }}
-                                                            />
-                                                        ) : log.typeEvenement?.libelle
-                                                    }
-                                                </td>
-                                                <td>
-                                                    {
-                                                        !log.new ? log.modifUser : null
-                                                    }
-                                                </td>
-                                                <td>
-                                                    {
-                                                        log.new || log.edit ? (
-                                                            <Tabou2TextForm
-                                                                type="text"
-                                                                value={log.description}
-                                                                style={{borderRadius: "4px"}}
-                                                                onChange={(e) => {
-                                                                    changeLog({...log, description: e.target.value});
-                                                                }}
-                                                                placeholder={log.description || props.i18n(props.messages, "tabou2.logsModal.notePlaceholder")} />
-
-                                                        ) : log.description
-                                                    }
-                                                </td>
-                                                {
-                                                    readOnly ? null :
-                                                        (<td>
-                                                            {log.new || log.edit ? (
-                                                                <Button
-                                                                    tooltip={props.i18n(props.messages, "tabou2.save")}
-                                                                    disabled={!log.typeEvenement || isEmpty(log.typeEvenement) || !log.description}
-                                                                    style={{ borderColor: "rgba(0,0,0,0)"}}
-                                                                    onClick={() => saveEvent(log)}>
-                                                                    <span style={{ color: "rgb(40, 167, 69)" }}>
-                                                                        <Glyphicon glyph="ok"/>
-                                                                    </span>
-                                                                </Button>) : null
-                                                            }
-                                                            {
-                                                                log.edit && !log.new ? (
-                                                                    <Button
-                                                                        tooltip={props.i18n(props.messages, "tabou2.cancel")}
-                                                                        style={{ borderColor: "rgba(0,0,0,0)"}}
-                                                                        onClick={() => cancelChange(log) }>
-                                                                        <span style={{color: "rgb(229,0,0)"}}>
-                                                                            <Glyphicon glyph="remove"/>
-                                                                        </span>
-                                                                    </Button>
-                                                                ) : null
-                                                            }
-                                                            {
-                                                                !log.new && !log.edit && !editionActivate.current ? (
-                                                                    <Button
-                                                                        tooltip={props.i18n(props.messages, "tabou2.change")}
-                                                                        style={{ borderColor: "rgba(0,0,0,0)"}}
-                                                                        onClick={() => changeLog({...log, edit: true}) }>
-                                                                        <span style={{color: "rgb(137,178,211)"}}>
-                                                                            <Glyphicon glyph="pencil"/>
-                                                                        </span>
-                                                                    </Button>
-                                                                ) : null
-                                                            }
-                                                            {
-                                                                log.new || !editionActivate.current && !log.systeme ? (
-                                                                    <Button
-                                                                        tooltip={props.i18n(props.messages, "tabou2.delete")}
-                                                                        style={{ borderColor: "rgba(0,0,0,0)"}}
-                                                                        onClick={() => deleteLog(log) }>
-                                                                        <span style={{color: "rgb(229,0,0)"}}>
-                                                                            <Glyphicon glyph="trash"/>
-                                                                        </span>
-                                                                    </Button>
-                                                                ) : null
-                                                            }
-                                                        </td>)
-                                                }
-                                            </tr>
-                                        </>
-                                    ))
-                                }
-                            </tbody>
-                        </Table>
-                    </Col>
-                </Row>
-            </Grid>
+            <Tabou2LogsTable
+                editionActivate={editionActivate}
+                disabledAdd={disabledAdd}
+                deleteLog={deleteLog}
+                logs={logs}
+                readOnly={readOnly}
+                sortField={sortField}
+                changeSort={changeSort}
+                saveEvent={saveEvent}
+                cancelChange={cancelChange}
+                changeLog={changeLog}
+                {...props}
+            />
         </ResizableModal>
     );
 }
